refactor: consumer parameters inputs
