+++ src/components/Publish/_validation.ts
@@ -3,7 +3,7 @@ import { MAX_DECIMALS } from '@utils/constants'
 import { getMaxDecimalsValidation } from '@utils/numbers'
 import * as Yup from 'yup'
 import { testLinks } from '../../@utils/yup'
-import { validationConsumerParameters } from '@components/@shared/FormInput/InputElement/ConsumerParameters/validationConsumerParameters'
+import { validationConsumerParameters } from '@components/@shared/FormInput/InputElement/ConsumerParameters/_validation'
 
 // TODO: conditional validation
 // e.g. when algo is selected, Docker image is required
